Meta module for the GLM class of models

The DiseasyModelGLM module implements common structure and functionality to GLM regression class of models beyond the model structure provided by DiseasyModelRegression.

Most notably, the model module implements the $fit_regression() and $get_prediction() methods using GLM.

diseasy includes two simple models that uses the DiseasyModelGLM module: DiseasyModelG0 and DiseasyModelG1 These models implements a constant predictor and a exponential model based on the previous 7 and 21 days of observations, respectively.

When making a custom GLM model, the subclass should implement the $update_formula() method. The $update_formula() method should update the formula based on the stratifications. If the model should flexibly adapt to different stratifications, this method should be implemented. See DiseasyModelG0 and DiseasyModelG1 for examples of how this can be done.

Value

A new instance of the DiseasyModelGLM, DiseasyModelG0 or DiseasyModelG1 R6 class.

Super classes

diseasy::DiseasyBaseModule -> diseasy::DiseasyModel -> diseasy::DiseasyModelRegression -> DiseasyModelGLM

Methods

Public methods

DiseasyModelGLM$clone()

Inherited methods

diseasy::DiseasyBaseModule$load_module()

diseasy::DiseasyBaseModule$set_moduleowner()

diseasy::DiseasyModel$get_data()

diseasy::DiseasyModelRegression$get_results()

diseasy::DiseasyModelRegression$initialize()

diseasy::DiseasyModelRegression$plot()

Method clone()

The objects of this class are cloneable with this method.

Usage

DiseasyModelGLM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

diseasy::DiseasyBaseModule -> diseasy::DiseasyModel -> diseasy::DiseasyModelRegression -> diseasy::DiseasyModelGLM -> diseasy::NA -> DiseasyModelG0

Methods

Public methods

DiseasyModelG0$new()

DiseasyModelG0$clone()

Inherited methods

diseasy::DiseasyBaseModule$load_module()

diseasy::DiseasyBaseModule$set_moduleowner()

diseasy::DiseasyModel$get_data()

diseasy::DiseasyModelRegression$get_results()

diseasy::DiseasyModelRegression$plot()

Method new()

Usage

DiseasyModelG0$new(...)

Arguments

...

parameters sent to DiseasyModelG_ R6 constructor

Returns

A new instance of the DiseasyModelG1 R6 class.

Method clone()

The objects of this class are cloneable with this method.

Usage

DiseasyModelG0$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

diseasy::DiseasyBaseModule -> diseasy::DiseasyModel -> diseasy::DiseasyModelRegression -> diseasy::DiseasyModelGLM -> diseasy::NA -> DiseasyModelG1

Methods

Public methods

DiseasyModelG1$new()

DiseasyModelG1$clone()

Inherited methods

diseasy::DiseasyBaseModule$load_module()

diseasy::DiseasyBaseModule$set_moduleowner()

diseasy::DiseasyModel$get_data()

diseasy::DiseasyModelRegression$get_results()

diseasy::DiseasyModelRegression$plot()

Method new()

Usage

DiseasyModelG1$new(...)

Arguments

...

parameters sent to DiseasyModelG_ R6 constructor

Returns

A new instance of the DiseasyModelG1 R6 class.

Method clone()

The objects of this class are cloneable with this method.

Usage

DiseasyModelG1$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

 model <- DiseasyModelG0$new()

 rm(model)