Base module for the framework

TODO

Value

A new instance of the DiseasyBaseModule

R6 class.

Active bindings

hash

(character)
Computes a hash value for the module. Useful for logging and caching. Read only.

Methods

Public methods

DiseasyBaseModule$new()

DiseasyBaseModule$set_moduleowner()

DiseasyBaseModule$load_module()

DiseasyBaseModule$finalize()

DiseasyBaseModule$clone()

Method new()

Creates a new instance of the DiseasyBaseModule R6 class. This module is typically not constructed directly but rather through derived classes.

Usage

DiseasyBaseModule$new(moduleowner = class(self)[1])

Arguments

moduleowner

(character)
The name of the moduleowner. Used when logging.

Method set_moduleowner()

Changes the "ownership" of the module. Used when logging.

Usage

DiseasyBaseModule$set_moduleowner(moduleowner)

Arguments

moduleowner

(character)
The name of the moduleowner.

Returns

NULL

Method load_module()

Loads a copy of the provided module into the module.

Usage

DiseasyBaseModule$load_module(module, clone = TRUE)

Arguments

module

(R6::R6Class instance)
This instance is cloned to the field with the same name as the class of the module

clone

(boolean)
Toggle whether or not the module should be cloned when loading. Default TRUE.

Details

The methods allows the setting of the internal module instances after the DiseasyBaseModule instance is created.

Returns

NULL

Method finalize()

Handles the clean-up of the class

Usage

DiseasyBaseModule$finalize()

Method clone()

The objects of this class are cloneable with this method.

Usage

DiseasyBaseModule$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

  # Normally, you would not want to create this module directly, but it is possible.
  base_module <- DiseasyBaseModule$new()
#> INFO  [2023-10-24 07:40:25.07705] DiseasyBaseModule: Module loaded

  rm(base_module)