dk_activity_units

This data set contains the "units" of activity that was developed for the Danish COVID-19 modelling efforts. Each "activity unit" specifies the amount of activity for the four arenas: Home, Work, School, Other. By combining these units of activity, a complete, granular picture of the societal activity is formed (see dk_reference_scenario for details).

Details

These units are developed to cover the proposed and realized scenarios of societal restrictions during the COVID-19 pandemic in Denmark. If other scenarios are to be tested, new "activity units" will (likely) have to be developed. We supply the script that generates these units with the package (see data-raw/dk-activity-units.R for details).

Author

Lasse Engbo Christiansen lsec@ssi.dk

Carsten Kirkeby ckir@sund.ku.dk

Frederik Plesner Lyngse frederik.lyngse@sund.ku.dk

Adam Mielke admi@dtu.dk

Rasmus Skytte Randl\u00F8v rske@dtu.dk