Observable handler

TODO

Value

A new instance of the DiseasyBaseModule

R6 class.

See also

Super class

diseasy::DiseasyBaseModule -> DiseasyObservables

Active bindings

case_definition

(character)
The set case_definition to get DiseasyObservables for. Read-only.

start_date

(Date)
The start date of the study period. Read-only.

end_date

(Date)
The end date of the study period. Read-only.

last_queryable_date

(Date)
The latest date that can be queried. Read-only.

ds

(Diseasystore*)
The currently loaded diseasystore which provides the features. Read-only.

available_observables

(character)
The currently available observables in the loaded diseasystore. Read-only.

slice_ts

(Date)
The timestamp to slice database on. Read-only.

conn

(DBIConnection)
The connection to the database on. Read-only.

Methods

Public methods

DiseasyObservables$new()

DiseasyObservables$set_case_definition()

DiseasyObservables$set_last_queryable_date()

DiseasyObservables$set_study_period()

DiseasyObservables$set_slice_ts()

DiseasyObservables$get_observation()

DiseasyObservables$describe()

DiseasyObservables$finalize()

DiseasyObservables$clone()

Inherited methods

diseasy::DiseasyBaseModule$load_module()

diseasy::DiseasyBaseModule$set_moduleowner()

Method new()

Creates a new instance of the DiseasyObservables R6 class.

Usage

DiseasyObservables$new(
  case_definition = NULL,
  start_date = NULL,
  end_date = NULL,
  last_queryable_date = NULL,
  conn = NULL,
  slice_ts = NULL,
  ...
)

Arguments

case_definition

(character)
A character string that controls which feature store to get data from.

start_date

(Date)
Study period start (default values for get_observation).

end_date

(Date)
Study period end (default values for get_observation).

last_queryable_date

(Date)
Enforce a limit on data that can be pulled (not after this date).

conn

(DBIConnection)
A database connection object (inherits from DBIConnection)

slice_ts

(Date or character)
Date to slice the database on. See SCDB::get_table()

...

parameters sent to DiseasyBaseModule R6 constructor.

Method set_case_definition()

Set the case definition to get DiseasyObservables for.

Usage

DiseasyObservables$set_case_definition(case_definition)

Arguments

case_definition

(character)
Text label of the disease to get DiseasyObservables for.
Must match case definition implemented in featurestore package.

Method set_last_queryable_date()

Enforce a limit on data that can be pulled.

Usage

DiseasyObservables$set_last_queryable_date(last_queryable_date)

Arguments

last_queryable_date

(Date)
DiseasyObservables module will not return data after this date.

Method set_study_period()

Set the (default) time period to get observations from.

Usage

DiseasyObservables$set_study_period(start_date, end_date)

Arguments

start_date

(Date)
Start date to get DiseasyObservables for (including).

end_date

(Date)
End date to get DiseasyObservables for (including).

Method set_slice_ts()

Set the slice_ts to get data for

Usage

DiseasyObservables$set_slice_ts(slice_ts)

Arguments

slice_ts

(Date or character)
Date to slice the database on

Method get_observation()

Retrieve an "observable" in the data set corresponding to the set case_definition.
By default, the internal values for start_date and end_date are used to return data, but these can be overwritten.
The results are cached for faster retrieval at subsequent calls.

Usage

DiseasyObservables$get_observation(
  observable,
  aggregation = NULL,
  start_date = self %.% start_date,
  end_date = self %.% end_date
)

Arguments

observable

(character)
The requested observable. Should follow the pattern 'n_*'.

aggregation

(list(quosures))
Default NULL. If given, expressions in aggregation evaluated to give the aggregation level.
Use rlang::quos(...) to specify aggregation.

start_date

(Date)
Start date to get DiseasyObservables for (including).

end_date

(Date)
End date to get DiseasyObservables for (including).

Returns

If the observable is found, the function returns the corresponding data at the aggregation level.
Otherwise, the function fails and lists the available DiseasyObservables for the case_definition.

Method describe()

Prints a human readable report of the internal state of the module.

Usage

DiseasyObservables$describe()

Method finalize()

Handles the clean-up of the class

Usage

DiseasyObservables$finalize()

Method clone()

The objects of this class are cloneable with this method.

Usage

DiseasyObservables$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

  # Create observables module using the Google COVID-19 data
  obs <- DiseasyObservables$new(case_definition = "Google COVID-19",
                                conn = DBI::dbConnect(RSQLite::SQLite()))
#> INFO  [2023-11-07 08:07:59.530492] DiseasyObservables: Module loaded
#> INFO  [2023-11-07 08:07:59.833116] DiseasyObservables: Case definition set to Google COVID-19 (Google COVID-19)

  # See available observables
  print(obs$available_observables)
#> [1] "n_population" "n_hospital"   "n_deaths"     "n_positive"   "n_icu"       
#> [6] "n_ventilator"

  # Get data for one observable
  if (FALSE) {
  obs$get_observation("n_hospital",
                      start_date = as.Date("2020-03-01"),
                      end_date = as.Date("2020-03-05"))
  }
  rm(obs)